+++ packages/better-auth/src/plugins/jwt/sign.ts
@@ -98,29 +98,6 @@ export async function generateExportedKeyPair(options?: JwtPluginOptions) {
 }
 
 /**
- * Backwards compatable version of signJwt
- *
- * @deprecated - prefer signJwt
- *
- * @param ctx - endpoint context
- * @param options - Jwt signing options. If not provided, uses the jwtPlugin options
- */
-export async function getJwtToken(
-	ctx: GenericEndpointContext,
-	options?: JwtPluginOptions,
-) {
-	if (!options) {
-		options = getJwtPlugin(ctx.context).options;
-	}
-
-	const payload = !options?.jwt?.definePayload
-		? ctx.context.session!.user
-		: await options?.jwt.definePayload(ctx.context.session!);
-
-	return await signJwt(ctx, payload, options);
-}
-
-/**
  * Creates a new JWK (JSON Web Key) on the database.
  */
 export async function createJwk(
